Change to a functional style immutable stack implementation
In order for the pop to both change the state and return a value, we had to
introduce another class 'Pair' that wraps both.
In the same way, push now returns the new Stack.

The stack implementation itself does not change and still uses side effects.

As the empty stack is 'null' all methods became static

The isEmpty method disapeared in favor of a direct comparision with the empty stack.